requirements.txt
pandas
numpy
networkx
node2vec
gensim
scikit-learn
joblib
shap
matplotlib
streamlit
azure-ai-inference
azure-identity